Remove extra bin
