+++ crates/dbg-swc/.gitignore
@@ -1,1 +1,2 @@
 *.output.js
+input.js
