+++ crates/dbg-swc/Cargo.toml
@@ -11,11 +11,6 @@ version = "0.12.2"
 bench = false
 name = "dbg-swc"
 
-[[bin]]
-bench = false
-name = "cmp-size"
-path = "src/cmp-size.rs"
-
 [dependencies]
 anyhow = "1.0.57"
 clap = { version = "3", features = ["derive"] }
