+++ crates/dbg-swc/src/minify/mod.rs
@@ -4,14 +4,16 @@ use anyhow::Result;
 use clap::Subcommand;
 use swc_common::SourceMap;
 
-use self::{compare::CompareCommand, ensure_size::EnsureSize};
+use self::{compare::CompareCommand, ensure_size::EnsureSize, reduce::ReduceCommand};
 
 mod compare;
 mod ensure_size;
+mod reduce;
 
 /// Minify a javascript file.
 #[derive(Debug, Subcommand)]
 pub enum MinifyCommand {
+    Reduce(ReduceCommand),
     Compare(CompareCommand),
     EnsureSize(EnsureSize),
 }
@@ -19,6 +21,7 @@ pub enum MinifyCommand {
 impl MinifyCommand {
     pub fn run(self, cm: Arc<SourceMap>) -> Result<()> {
         match self {
+            MinifyCommand::Reduce(cmd) => cmd.run(cm),
             MinifyCommand::EnsureSize(cmd) => cmd.run(cm),
             MinifyCommand::Compare(cmd) => cmd.run(cm),
         }
